# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechShop E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  role: UserRole;
  isActive?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserUpdate {
  name?: string;
  phone?: string;
  address?: string;
  isActive?: boolean;
  role?: UserUpdateRole;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
  createdAt?: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  images?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  sku?: string | null;
  isFeatured?: boolean;
  isActive?: boolean;
  soldCount?: number;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  originalPrice?: number;
  stock: number;
  imageUrl?: string;
  images?: string;
  categoryId: number;
  brand?: string;
  sku?: string;
  isFeatured?: boolean;
  isActive?: boolean;
}

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  stock?: number;
  imageUrl?: string;
  images?: string;
  categoryId?: number;
  brand?: string;
  sku?: string;
  isFeatured?: boolean;
  isActive?: boolean;
}

export type StockUpdateOperation = typeof StockUpdateOperation[keyof typeof StockUpdateOperation];


export const StockUpdateOperation = {
  set: 'set',
  add: 'add',
  subtract: 'subtract',
} as const;

export interface StockUpdate {
  stock: number;
  operation?: StockUpdateOperation;
}

export interface ProductVariant {
  id: number;
  productId: number;
  name: string;
  value: string;
  priceModifier: number;
  stock: number;
  /** @nullable */
  sku?: string | null;
  isDefault: boolean;
  createdAt: string;
}

export interface ProductVariantInput {
  name: string;
  value: string;
  priceModifier?: number;
  stock?: number;
  sku?: string;
  isDefault?: boolean;
}

export interface ProductVariantUpdate {
  name?: string;
  value?: string;
  priceModifier?: number;
  stock?: number;
  sku?: string;
  isDefault?: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CartItem {
  id: number;
  productId: number;
  productName?: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
  subtotal?: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
  subtotal: number;
}

export interface DeliveryArea {
  wilayat: string;
  wilayatAr?: string;
  deliveryChargeOmr: number;
  estimatedDays: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  cod: 'cod',
  online: 'online',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  trackingId?: string;
  status: OrderStatus;
  total: number;
  deliveryCharge?: number;
  paymentMethod?: OrderPaymentMethod;
  deliveryArea?: string;
  codDeliveryChargePaid?: boolean;
  shippingAddress?: string;
  phone?: string;
  /** @nullable */
  notes?: string | null;
  items?: OrderItem[];
  createdAt: string;
  updatedAt?: string;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cod: 'cod',
  online: 'online',
} as const;

export interface OrderInput {
  shippingAddress: string;
  phone: string;
  deliveryArea: string;
  paymentMethod: OrderInputPaymentMethod;
  notes?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  sortOrder?: number;
  isActive: boolean;
  createdAt?: string;
}

export interface BannerInput {
  title: string;
  subtitle?: string;
  imageUrl: string;
  linkUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface BannerUpdate {
  title?: string;
  subtitle?: string;
  imageUrl?: string;
  linkUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface CategoryRevenue {
  categoryName: string;
  revenue: number;
  orderCount: number;
}

export interface AdminStats {
  totalUsers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  lowStockCount: number;
  todayOrders: number;
  todayRevenue: number;
  ordersByStatus?: StatusCount[];
  revenueByCategory?: CategoryRevenue[];
}

export type ListUsersParams = {
page?: number;
limit?: number;
search?: string;
};

export type ListProductsParams = {
page?: number;
limit?: number;
search?: string;
categoryId?: number;
minPrice?: number;
maxPrice?: number;
sortBy?: string;
featured?: boolean;
};

export type ListOrdersParams = {
page?: number;
limit?: number;
status?: string;
userId?: number;
};

export type GetLowStockProductsParams = {
threshold?: number;
};