# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechShop E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, user sees own)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: userId
          schema: { type: integer }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track/{trackingId}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by tracking ID
      parameters:
        - in: path
          name: trackingId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Banners
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, price, subtotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: ["string", "null"] }
        quantity: { type: integer }
        price: { type: number }
        subtotal: { type: number }
...
    Order:
      type: object
      required: [id, userId, status, total, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userName: { type: ["string", "null"] }
        userEmail: { type: ["string", "null"] }
        trackingId: { type: string }
        status: { type: string, enum: [pending, confirmed, processing, shipped, delivered, cancelled] }
        total: { type: number }
        deliveryCharge: { type: number }
        paymentMethod: { type: string, enum: [cod, online] }
        deliveryArea: { type: string }
        codDeliveryChargePaid: { type: boolean }
        shippingAddress: { type: string }
        phone: { type: string }
        notes: { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt: { type: string }
        updatedAt: { type: string }
    OrderInput:
      type: object
      required: [shippingAddress, phone, deliveryArea, paymentMethod]
      properties:
        shippingAddress: { type: string }
        phone: { type: string }
        deliveryArea: { type: string }
        paymentMethod: { type: string, enum: [cod, online] }
        notes: { type: string }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, confirmed, processing, shipped, delivered, cancelled] }
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...