# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechShop E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  trackingId?: string;
  status: OrderStatus;
  total: number;
  deliveryCharge?: number;
  paymentMethod?: OrderPaymentMethod;
  deliveryArea?: string;
  codDeliveryChargePaid?: boolean;
  shippingAddress?: string;
  phone?: string;
  /** @nullable */
  notes?: string | null;
  items?: OrderItem[];
  createdAt: string;
  updatedAt?: string;
}